requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
data-persistence-repository==0.2.1
exceptiongroup==1.2.0
fastapi==0.108.0
greenlet==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.3
pytest-asyncio==0.23.2
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.32.0.post1
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
